+++ packages/engine/src/Actions.js
@@ -40,6 +40,7 @@ class Actions {
         });
         responses[action.id] = response;
       } catch (error) {
+        responses[action.id] = error;
         throw {
           error,
           action,
@@ -54,12 +55,10 @@ class Actions {
     try {
       await this.callActionLoop({ actions, arrayIndices, block, event, responses });
     } catch (error) {
-      responses[error.action.id] = error.error;
       console.error(error);
       try {
         await this.callActionLoop({ actions: catchActions, arrayIndices, block, event, responses });
       } catch (errorCatch) {
-        responses[errorCatch.action.id] = errorCatch.error;
         console.error(errorCatch);
         return {
           blockId: block.blockId,
@@ -132,12 +131,24 @@ class Actions {
         params: parsedAction.params,
       });
     } catch (error) {
+      responses[action.id] = { error, index, type: action.type };
+      const { output: parsedMessages, errors: parserErrors } = this.context.parser.parse({
+        actions: responses,
+        event,
+        arrayIndices,
+        input: action.messages,
+        location: block.blockId,
+      });
+      if (parserErrors.length > 0) {
+        // this condition is very unlikely since parser errors usually occur in the first parse.
+        throw { error: parserErrors[0], type: action.type, index };
+      }
       closeLoading();
       this.displayMessage({
-        defaultMessage: error.lowdefyMessage || 'Action unsuccessful',
+        defaultMessage: error.message,
         duration: 6,
         hideExplicitly: true,
-        message: messages.error,
+        message: (parsedMessages || {}).error,
         status: 'error',
       });
       throw {
@@ -157,7 +168,6 @@ class Actions {
 
   displayMessage({ defaultMessage, duration, hideExplicitly, message, status }) {
     let close = () => undefined;
-
     if ((hideExplicitly && message !== false) || (!hideExplicitly && !type.isNone(message))) {
       close = this.context.lowdefy.displayMessage({
         content: type.isString(message) ? message : defaultMessage,